refactor(lotto) : 패키지 위치 조정
LottoNumber dto패키지 에서 out
LottoWinningNumber dto 패키지 에서 out

issue #408
