+++ src/main/java/lotto/domain/LottoChecker.java
@@ -1,9 +1,7 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoNumber;
 import lotto.domain.dto.LottoRank;
 import lotto.domain.dto.LottoResult;
-import lotto.domain.dto.LottoWinningNumber;
 
 import java.util.Arrays;
 import java.util.List;
