+++ src/test/java/lotto/domain/LottoBuyerTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
